backend/requirements.txt
amqp==5.2.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-sass-processor==1.4.1
django-timezone-field==7.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
flower==2.0.1
h11==0.14.0
humanize==4.11.0
idna==3.10
kombu==5.4.2
libsass==0.23.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.1.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
tornado==6.4.1
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
